+++ opendmi/src/format/text/handlers.c
@@ -14,34 +14,13 @@
 #include <inttypes.h>
 #include <assert.h>
 
-#ifdef ENABLE_CURSES
-#   if __has_include(<ncurses/ncurses.h>)
-#       include <ncurses/ncurses.h>
-#   elif __has_include(<ncurses/curses.h>)
-#       include <ncurses/curses.h>
-#   elif __has_include(<ncurses.h>)
-#       include <ncurses.h>
-#   elif __has_include(<curses.h>)
-#       include <curses.h>
-#   endif
-#   include <term.h>
-#endif // ENABLE_CURSES
-
 #include <opendmi/context.h>
 #include <opendmi/error.h>
 #include <opendmi/utils.h>
+#include <opendmi/utils/tty.h>
 
 #include <opendmi/format/text/handlers.h>
 #include <opendmi/format/text/helpers.h>
-
-#define COLOR_NONE (-1)
-#define COLOR_GREY 8
-
-static void dmi_tprintf(
-        dmi_text_session_t *session,
-        int color,
-        const char *format,
-        ...);
 
 void *dmi_text_initialize(dmi_context_t *context, FILE *stream)
 {
@@ -79,12 +58,14 @@ bool dmi_text_entry(dmi_text_session_t *session)
     if (entity_count == 0)
         entity_count = context->registry->count;
 
-    dmi_tprintf(session, COLOR_NONE, "SMBIOS %s present\n", version);
-    dmi_tprintf(session, COLOR_NONE, "SMBIOS vendor: %s\n", context->vendor_name ? context->vendor_name : "unknown");
-    dmi_tprintf(session, COLOR_NONE, "%zu structures occupying %zu bytes\n",
-                entity_count, context->table_area_size);
-    dmi_tprintf(session, COLOR_NONE, "Table at 0x%" PRIx64 "\n", context->table_area_addr);
-    dmi_tprintf(session, COLOR_NONE, "\n");
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "SMBIOS %s present\n", version);
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "SMBIOS vendor: %s\n",
+                    context->vendor_name ? context->vendor_name : "unknown");
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "%zu structures occupying %zu bytes\n",
+                    entity_count, context->table_area_size);
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "Table at 0x%" PRIx64 "\n",
+                    context->table_area_addr);
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "\n");
 
     dmi_free(version);
 
@@ -96,11 +77,11 @@ bool dmi_text_entity_start(dmi_text_session_t *session, const dmi_entity_t *enti
     assert(session != nullptr);
     assert(entity != nullptr);
 
-    dmi_tprintf(session, COLOR_YELLOW, "Handle 0x%04hX, DMI type %d, %zu bytes\n",
-                dmi_entity_handle(entity),
-                dmi_entity_type(entity),
-                entity->total_length);
-    dmi_tprintf(session, COLOR_YELLOW, "%s\n", dmi_entity_name(entity));
+    dmi_text_printf(session, DMI_TTY_COLOR_YELLOW, "Handle 0x%04hX, DMI type %d, %zu bytes\n",
+                    dmi_entity_handle(entity),
+                    dmi_entity_type(entity),
+                    entity->total_length);
+    dmi_text_printf(session, DMI_TTY_COLOR_YELLOW, "%s\n", dmi_entity_name(entity));
 
     return true;
 }
@@ -117,7 +98,7 @@ bool dmi_text_entity_attr(
     assert(value != nullptr);
 
     // Print attribute name
-    dmi_tprintf(session, COLOR_NONE, "\t%s: ", attr->params.name);
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "\t%s: ", attr->params.name);
 
     // Print attribute value
     if (not dmi_member_is_present(attr->counter)) {
@@ -147,10 +128,10 @@ void dmi_text_entity_attr_array(
     size_t count = dmi_member_value(info, attr->counter, size_t);
     const dmi_data_t *ptr = *(const dmi_data_t **)value;
 
-    dmi_tprintf(session, COLOR_NONE, "%zu items\n", count);
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "%zu items\n", count);
 
     for (size_t i = 0; i < count; i++, ptr += attr->value.size) {
-        dmi_tprintf(session, COLOR_NONE, "\t\t%zu: ", i);
+        dmi_text_printf(session, DMI_TTY_COLOR_NONE, "\t\t%zu: ", i);
 
         if (attr->type == DMI_ATTRIBUTE_TYPE_STRUCT) {
             dmi_text_entity_attr_struct(session, attr, ptr);
@@ -180,11 +161,11 @@ void dmi_text_entity_attr_struct(
 
     const dmi_attribute_t *child_attr = nullptr;
 
-    dmi_tprintf(session, COLOR_NONE, "\n");
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "\n");
     for (child_attr = attr->params.attrs; child_attr->params.name; child_attr++) {
         const dmi_data_t *ptr = dmi_member_ptr(value, child_attr->value, dmi_data_t);
 
-        dmi_tprintf(session, COLOR_NONE, "\t\t\t%s: ", child_attr->params.name);
+        dmi_text_printf(session, DMI_TTY_COLOR_NONE, "\t\t\t%s: ", child_attr->params.name);
         dmi_text_entity_attr_value(session, child_attr, ptr, nullptr);
     }
 }
@@ -202,27 +183,27 @@ void dmi_text_entity_attr_value(
     char *text;
 
     if (dmi_attribute_is_unspecified(attr, value)) {
-        dmi_tprintf(session, COLOR_GREY, "<unspecified>\n");
+        dmi_text_printf(session, DMI_TTY_COLOR_GREY, "<unspecified>\n");
         return;
     }
     if (dmi_attribute_is_unknown(attr, value)) {
-        dmi_tprintf(session, COLOR_CYAN, "<unknown>\n");
+        dmi_text_printf(session, DMI_TTY_COLOR_MAROON, "<unknown>\n");
         return;
     }
 
     text = dmi_attribute_format(attr, value, true);
     if (text == nullptr) {
-        dmi_tprintf(session, COLOR_RED, "<error>\n");
+        dmi_text_printf(session, DMI_TTY_COLOR_RED, "<error>\n");
         return;
     }
 
     if (attr->params.unit)
-        dmi_tprintf(session, COLOR_NONE, "%s %s", text, attr->params.unit);
+        dmi_text_printf(session, DMI_TTY_COLOR_NONE, "%s %s", text, attr->params.unit);
     else
-        dmi_tprintf(session, COLOR_NONE, "%s", text);
+        dmi_text_printf(session, DMI_TTY_COLOR_NONE, "%s", text);
 
     if (descr != nullptr)
-        dmi_tprintf(session, COLOR_NONE, " - %s", descr);
+        dmi_text_printf(session, DMI_TTY_COLOR_NONE, " - %s", descr);
 
     fputc('\n', session->stream);
 
@@ -249,7 +230,9 @@ void dmi_text_entity_attr_set(
             continue;
 
         bool flag = mask & (1 << i);
-        dmi_tprintf(session, COLOR_NONE, "\t\t%s: %s\n", name, flag ? "yes" : "no");
+
+        dmi_text_printf(session, DMI_TTY_COLOR_NONE, "\t\t%s: %s\n",
+                        name, flag ? "yes" : "no");
     }
 }
 
@@ -258,7 +241,7 @@ bool dmi_text_entity_data(dmi_text_session_t *session, const dmi_entity_t *entit
     assert(session != nullptr);
     assert(entity != nullptr);
 
-    dmi_tprintf(session, COLOR_NONE, "\tHeader and data:\n");
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "\tHeader and data:\n");
     dmi_text_hex_data(session, entity->data, entity->body_length);
 
     return true;
@@ -272,11 +255,11 @@ bool dmi_text_entity_strings(dmi_text_session_t *session, const dmi_entity_t *en
     if (entity->string_count == 0)
         return true;
 
-    dmi_tprintf(session, COLOR_NONE, "\tStrings:\n");
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "\tStrings:\n");
 
     for (dmi_string_t i = 1; i <= entity->string_count; i++) {
         const char *str = dmi_entity_string_ex(entity, i, true);
-        dmi_tprintf(session, COLOR_NONE, "\t\t%u: \"%s\"\n", i, str);
+        dmi_text_printf(session, DMI_TTY_COLOR_NONE, "\t\t%u: \"%s\"\n", i, str);
     }
 
     return true;
@@ -289,7 +272,7 @@ bool dmi_text_entity_end(dmi_text_session_t *session, const dmi_entity_t *entity
 
     dmi_unused(entity);
 
-    dmi_tprintf(session, COLOR_NONE, "\n");
+    dmi_text_printf(session, DMI_TTY_COLOR_NONE, "\n");
 
     return true;
 }
@@ -299,27 +282,4 @@ void dmi_text_finalize(dmi_text_session_t *session)
     assert(session != nullptr);
 
     dmi_free(session);
-}
-
-static void dmi_tprintf(
-        dmi_text_session_t *session,
-        int color,
-        const char *format,
-        ...)
-{
-    va_list args;
-
-#ifdef ENABLE_CURSES
-    if (session->is_tty and (color >= 0))
-        tputs(tparm(tigetstr("setaf"), color), 1, putchar);
-#endif // ENABLE_CURSES
-
-    va_start(args, format);
-    vfprintf(session->stream, format, args);
-    va_end(args);
-
-#ifdef ENABLE_CURSES
-    if (session->is_tty and (color >= 0))
-        tputs(tparm(tigetstr("sgr0")), 1, putchar);
-#endif // ENABLE_CURSES
 }